rename from input.modulse.css to form.module.css, and from start to flex-start.
