+++ src/page-modules/contact/components/form/form.module.css
@@ -441,7 +441,7 @@
   gap: var(--spacings-small);
 
   display: flex;
-  justify-content: start;
+  justify-content: flex-start;
   align-items: center;
   padding: var(--spacings-small);
 }
