+++ src/page-modules/contact/components/form/searchable-select/searchable-select.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { MonoIcon } from '@atb/components/icon';
 import ErrorMessage from '../error-message';
 import { andIf } from '@atb/utils/css';
-import style from '../input.module.css';
+import style from '../form.module.css';
 import {
   Key,
   Button,
